Cast NSInteger to long for formatting as %ld
In https://crbug.com/747643, a new version of Clang produces warnings
when attempting to format NSInteger as %zd (which is normally used for
ssize_t). Acquiesce by casting NSInteger to long and formatting using
%ld.

In the process, I audited all occurrences of %zd in GTM and found that
the remainder were being used to format size_t, which ought to use %zu,
so I’ve corrected those.

I also audited the existing uses of %zu and found them all to be
correct.

+++ iPhone/GTMUIImage+Resize.m
@@ -152,7 +152,7 @@ GTM_INLINE CGSize swapWidthAndHeight(CGSize size) {
       break;
 
     default:
-      _GTMDevAssert(false, @"Invalid orientation %zd", orientation);
+      _GTMDevAssert(false, @"Invalid orientation %ld", (long)orientation);
       return nil;
   }
